requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-cloud-firestore>=2.19.0
google-cloud-aiplatform>=1.60.0
google-genai>=1.0.0
uvloop>=0.19.0
jinja2>=3.1.0
